+++ NEWS.md
@@ -2,7 +2,12 @@
 * Compatibility fix with `ggplot` (>= 4.0.0)
 * Correct sensitivity of `irep()` to sort order (@jpryby, #226)
 * Improve reliability of NM-TRAN/NMTRAN message detextion (rfaelens #200)
-
+* Additional features (sbihorel #218)
+  * the addition of the `tad` variable type for `Time after dose`
+  * new tad-related goodness of fit plots
+  * new functions to set and get variables and units (including new label_units data slots to support label and unit use for simulated data)
+  * automated detection and use of variable label and units in scatterplots and vpc plots
+  * new function to apply SAS-like data formatting to convert numerical data into text-based factors
 
 # xpose 0.4.19
 * Fixed bug in condition number when Eigen values outputted on multiple records in .lst file (@billdenney & @marianklose, #128)